requirements.txt
kafka-python==1.3.3
pymongo==3.4.0
python-dateutil==2.6.0
